add some services and add vue as a lib

+++ src/main/java/ru/trainee/adminPanel/repository/BannerRepository.java
@@ -3,5 +3,5 @@ package ru.trainee.adminPanel.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.trainee.adminPanel.model.Banner;
 
-public interface BannerRepository extends JpaRepository<Banner, Long> {
+public interface BannerRepository extends JpaRepository<Banner, Integer> {
 }
